requirements.txt
certifi==2018.11.29
chardet==3.0.4
dj-database-url==0.5.0
Django==1.11.3
django-ckeditor==5.6.1
django-heroku==0.3.1
django-js-asset==1.1.0
gunicorn==19.9.0
idna==2.7
Pillow==5.3.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pytz==2018.7
requests==2.19.1
urllib3==1.23
whitenoise==4.1.2
